RoutesManagementService/requirements.txt
django>=5.0,<5.3
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
confluent-kafka>=2.3,<2.5
redis>=5.0,<5.1
python-dotenv>=1.0,<1.1
djangorestframework>=3.14,<3.15